requirements.txt
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
coverage==7.3.4
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
filelock==3.13.1
fonttools==4.44.3
fsspec==2023.10.0
future==0.18.2
fvcore==0.1.5.post20221221
httplib2==0.20.2
idna==3.3
imageio==2.33.0
importlib-metadata==4.6.4
iniconfig==2.0.0
iopath==0.1.10
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
openslide-python==1.3.1
packaging==23.2
pandas==2.1.3
paramiko==2.9.3
patsy==0.5.4
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.3.0
portalocker==2.8.2
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycore==17.4.17
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.4.3
pytest-cov==4.1.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scikit-learn==1.3.2
scipy==1.11.4
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
statsmodels==0.14.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
yacs==0.1.8
zipp==1.0.0
